chore:  tests for getModulePath and getTextFromFilePath
