+++ src/index.js
@@ -1,8 +1,6 @@
 /* eslint no-console:0, global-require:0, import/no-dynamic-require:0 */
 /* eslint complexity: [1, 13] */
-import fs from 'fs';
 import path from 'path';
-import requireRelative from 'require-relative';
 import prettyFormat from 'pretty-format';
 import { oneLine, stripIndent } from 'common-tags';
 import indentString from 'indent-string';
@@ -11,7 +9,9 @@ import merge from 'lodash.merge';
 import {
   getESLintCLIEngine,
   getOptionsForFormatting,
-  requireModule
+  requireModule,
+  getTextFromFilePath,
+  getModulePath
 } from './utils';
 
 const logger = getLogger({ prefix: 'prettier-eslint' });
@@ -56,7 +56,7 @@ function format(options) {
     prettierLast,
     fallbackPrettierOptions
   } = options;
-
+  
   const eslintConfig = merge(
     {},
     options.eslintConfig,
@@ -201,25 +201,7 @@ function createEslintFix(eslintConfig, eslintPath) {
   };
 }
 
-function getTextFromFilePath(filePath) {
-  try {
-    logger.trace(
-      oneLine`
-        attempting fs.readFileSync to get
-        the text for file at "${filePath}"
-      `
-    );
-    return fs.readFileSync(filePath, 'utf8');
-  } catch (error) {
-    logger.error(
-      oneLine`
-        failed to get the text to format
-        from the given filePath: "${filePath}"
-      `
-    );
-    throw error;
-  }
-}
+
 
 function getESLintConfig(filePath, eslintPath) {
   const eslintOptions = {};
@@ -261,21 +243,6 @@ function getPrettierConfig(filePath, prettierPath) {
   );
 }
 
-function getModulePath(filePath = __filename, moduleName) {
-  try {
-    return requireRelative.resolve(moduleName, filePath);
-  } catch (error) {
-    logger.debug(
-      oneLine`
-        There was a problem finding the ${moduleName}
-        module. Using prettier-eslint's version.
-      `,
-      error.message,
-      error.stack
-    );
-    return require.resolve(moduleName);
-  }
-}
 
 function getDefaultLogLevel() {
   return process.env.LOG_LEVEL || 'warn';
