+++ tests/fixtures/utils/text.js
@@ -1,0 +1,1 @@
+var foo = 'bar'
